Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
razorpay
openai
python-dotenv
fastapi
uvicorn
sqlalchemy
python-jose
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1